% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{abind_first}
\alias{abind_first}
\title{Bind Two Arrays Along the First Dimension}
\usage{
abind_first(a, b)
}
\arguments{
\item{a}{First array}

\item{b}{Second array (must match dimensions of \code{a} except along dim 1)}
}
\value{
Combined array
}
\description{
Concatenates two arrays along the first dimension (samples axis).
}
\keyword{internal}
