% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ag_device.R
\name{ag_default_device}
\alias{ag_default_device}
\title{Return the current default compute device}
\usage{
ag_default_device()
}
\value{
\code{"cpu"} or \code{"gpu"}
}
\description{
Return the current default compute device
}
