% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ag_device.R
\name{ag_default_dtype}
\alias{ag_default_dtype}
\title{Return the current default dtype for GPU operations}
\usage{
ag_default_dtype()
}
\value{
\code{"f32"}, \code{"f16"}, or \code{"bf16"}
}
\description{
Return the current default dtype for GPU operations
}
