% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ag_layers.R
\name{ag_eval}
\alias{ag_eval}
\title{Switch a layer or sequential model to eval mode}
\usage{
ag_eval(model)
}
\arguments{
\item{model}{An ag_sequential, ag_batch_norm, or ag_dropout layer}
}
\value{
The model/layer (invisibly)
}
\description{
Switch a layer or sequential model to eval mode
}
