% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autograd.R
\name{ag_mean}
\alias{ag_mean}
\title{Mean of elements (or along a dim)}
\usage{
ag_mean(x, dim = NULL, keepdim = FALSE)
}
\arguments{
\item{x}{ag_tensor}

\item{dim}{NULL (all), 1 (row-wise), or 2 (col-wise)}

\item{keepdim}{Logical}
}
\value{
ag_tensor
}
\description{
Mean of elements (or along a dim)
}
