% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autograd.R
\name{ag_mse_loss}
\alias{ag_mse_loss}
\title{Mean Squared Error loss}
\usage{
ag_mse_loss(pred, target)
}
\arguments{
\item{pred}{ag_tensor [units, batch_size]}

\item{target}{ag_tensor or matrix [units, batch_size]}
}
\value{
scalar ag_tensor
}
\description{
Mean Squared Error loss
}
