% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autograd.R
\name{ag_mul}
\alias{ag_mul}
\title{Element-wise multiplication}
\usage{
ag_mul(A, B)
}
\arguments{
\item{A}{ag_tensor or numeric matrix}

\item{B}{ag_tensor or numeric matrix}
}
\value{
ag_tensor
}
\description{
Element-wise multiplication
}
