% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autograd.R
\name{ag_pow}
\alias{ag_pow}
\title{Element-wise power}
\usage{
ag_pow(x, p)
}
\arguments{
\item{x}{ag_tensor}

\item{p}{Numeric exponent (scalar, not tracked for gradients)}
}
\value{
ag_tensor
}
\description{
Element-wise power
}
