% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autograd.R
\name{ag_reshape}
\alias{ag_reshape}
\title{Reshape tensor}
\usage{
ag_reshape(x, nrow, ncol)
}
\arguments{
\item{x}{ag_tensor}

\item{nrow}{New number of rows (use -1 to infer)}

\item{ncol}{New number of columns (use -1 to infer)}
}
\value{
ag_tensor with new shape, same data
}
\description{
Reshape tensor
}
