% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autograd.R
\name{ag_scale}
\alias{ag_scale}
\title{Scale tensor by a scalar constant}
\usage{
ag_scale(x, scalar)
}
\arguments{
\item{x}{ag_tensor}

\item{scalar}{Numeric scalar (not tracked for gradients)}
}
\value{
ag_tensor
}
\description{
Scale tensor by a scalar constant
}
