% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autograd.R
\name{ag_sub}
\alias{ag_sub}
\title{Element-wise subtraction}
\usage{
ag_sub(A, B)
}
\arguments{
\item{A}{ag_tensor or numeric matrix}

\item{B}{ag_tensor or numeric matrix}
}
\value{
ag_tensor
}
\description{
Element-wise subtraction
}
