% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ag_trace.R
\name{ag_trace_data_report}
\alias{ag_trace_data_report}
\title{Report traced reads of `ag_tensor$data`}
\usage{
ag_trace_data_report(n = 40L)
}
\arguments{
\item{n}{Maximum number of call sites to print.}
}
\value{
A data frame of sites and counts, invisibly.
}
\description{
Report traced reads of `ag_tensor$data`
}
\keyword{internal}
