% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autograd.R
\name{ag_transpose}
\alias{ag_transpose}
\title{Transpose a tensor}
\usage{
ag_transpose(x)
}
\arguments{
\item{x}{ag_tensor}
}
\value{
ag_tensor with rows and columns swapped
}
\description{
Transpose a tensor
}
