% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quants.R
\name{dequantize_row_q2_K}
\alias{dequantize_row_q2_K}
\alias{dequantize_row_q3_K}
\alias{dequantize_row_q4_K}
\alias{dequantize_row_q5_K}
\alias{dequantize_row_q6_K}
\alias{dequantize_row_q8_K}
\title{Dequantize Row (K-quants)}
\usage{
dequantize_row_q2_K(raw_data, n_elements)

dequantize_row_q3_K(raw_data, n_elements)

dequantize_row_q4_K(raw_data, n_elements)

dequantize_row_q5_K(raw_data, n_elements)

dequantize_row_q6_K(raw_data, n_elements)

dequantize_row_q8_K(raw_data, n_elements)
}
\arguments{
\item{raw_data}{Raw vector containing quantized data}

\item{n_elements}{Number of elements to dequantize}
}
\value{
Numeric vector of dequantized values
}
\description{
Converts K-quant quantized data back to float values.
K-quants (q2_K through q8_K) provide better quality/size tradeoffs.
}
\seealso{
Other quantization: 
\code{\link{dequantize_row_iq2_xxs}()},
\code{\link{dequantize_row_mxfp4}()},
\code{\link{dequantize_row_nvfp4}()},
\code{\link{dequantize_row_q1_0}()},
\code{\link{dequantize_row_q4_0}()},
\code{\link{dequantize_row_tq1_0}()},
\code{\link{ggml_quant_block_info}()},
\code{\link{iq2xs_free_impl}()},
\code{\link{iq2xs_init_impl}()},
\code{\link{iq3xs_free_impl}()},
\code{\link{iq3xs_init_impl}()},
\code{\link{quantize_iq2_xxs}()},
\code{\link{quantize_mxfp4}()},
\code{\link{quantize_nvfp4}()},
\code{\link{quantize_q1_0}()},
\code{\link{quantize_q2_K}()},
\code{\link{quantize_q4_0}()},
\code{\link{quantize_row_iq3_xxs_ref}()},
\code{\link{quantize_row_mxfp4_ref}()},
\code{\link{quantize_row_nvfp4_ref}()},
\code{\link{quantize_row_q1_0_ref}()},
\code{\link{quantize_row_q2_K_ref}()},
\code{\link{quantize_row_q4_0_ref}()},
\code{\link{quantize_row_tq1_0_ref}()},
\code{\link{quantize_tq1_0}()}
}
\concept{quantization}
