% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{ggml_abort_is_r_enabled}
\alias{ggml_abort_is_r_enabled}
\title{Check if R Abort Handler is Enabled}
\usage{
ggml_abort_is_r_enabled()
}
\value{
Logical indicating if R-compatible abort handling is active
}
\description{
Check if R Abort Handler is Enabled
}
\seealso{
Other logging: 
\code{\link{ggml_log_is_r_enabled}()},
\code{\link{ggml_log_set_default}()},
\code{\link{ggml_log_set_r}()},
\code{\link{ggml_set_abort_callback_default}()},
\code{\link{ggml_set_abort_callback_r}()}
}
\concept{logging}
