% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{ggml_abs_inplace}
\alias{ggml_abs_inplace}
\title{Absolute Value In-place (Graph)}
\usage{
ggml_abs_inplace(ctx, a)
}
\arguments{
\item{ctx}{GGML context}

\item{a}{Input tensor (will be modified in-place)}
}
\value{
View of tensor a with absolute values
}
\description{
Creates a graph node for in-place element-wise absolute value.
}
