% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{ggml_add_rel_pos}
\alias{ggml_add_rel_pos}
\title{Add Relative Position Bias (Graph)}
\usage{
ggml_add_rel_pos(ctx, a, pw, ph)
}
\arguments{
\item{ctx}{GGML context}

\item{a}{Input tensor}

\item{pw}{Width relative-position tensor}

\item{ph}{Height relative-position tensor}
}
\value{
Tensor with added relative-position bias
}
\description{
Adds width and height relative-position bias to \code{a}.
}
