% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{ggml_arange}
\alias{ggml_arange}
\title{Arange (Graph)}
\usage{
ggml_arange(ctx, start, stop, step = 1)
}
\arguments{
\item{ctx}{GGML context}

\item{start}{Start value (inclusive)}

\item{stop}{Stop value (exclusive)}

\item{step}{Step between values (default 1)}
}
\value{
1D F32 tensor
}
\description{
Creates a 1D F32 tensor with values from \code{start} (inclusive) to
\code{stop} (exclusive) in steps of \code{step}.
}
