% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{ggml_backend_buffer_free}
\alias{ggml_backend_buffer_free}
\title{Free Backend Buffer}
\usage{
ggml_backend_buffer_free(buffer)
}
\arguments{
\item{buffer}{Backend buffer object}
}
\value{
No return value, called for side effects
}
\description{
Frees a backend buffer and all associated memory.
}
