% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{ggml_backend_buffer_get_size}
\alias{ggml_backend_buffer_get_size}
\title{Get Backend Buffer Size}
\usage{
ggml_backend_buffer_get_size(buffer)
}
\arguments{
\item{buffer}{Backend buffer object}
}
\value{
Size in bytes
}
\description{
Returns the total size of a backend buffer.
}
