% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{ggml_backend_buffer_name}
\alias{ggml_backend_buffer_name}
\title{Get Backend Buffer Name}
\usage{
ggml_backend_buffer_name(buffer)
}
\arguments{
\item{buffer}{Backend buffer object}
}
\value{
Character string with buffer name
}
\description{
Returns the name/type of a backend buffer.
}
