% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{ggml_backend_cpu_set_n_threads}
\alias{ggml_backend_cpu_set_n_threads}
\title{Set CPU Backend Threads}
\usage{
ggml_backend_cpu_set_n_threads(backend, n_threads)
}
\arguments{
\item{backend}{CPU backend pointer}

\item{n_threads}{Number of threads}
}
\value{
NULL invisibly
}
\description{
Sets the number of threads for CPU backend computation.
}
