% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{ggml_backend_free}
\alias{ggml_backend_free}
\title{Free Backend}
\usage{
ggml_backend_free(backend)
}
\arguments{
\item{backend}{Backend pointer}
}
\value{
NULL invisibly
}
\description{
Releases resources associated with a backend.
}
