% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduler.R
\name{ggml_backend_sched_get_n_backends}
\alias{ggml_backend_sched_get_n_backends}
\title{Get number of backends in scheduler}
\usage{
ggml_backend_sched_get_n_backends(sched)
}
\arguments{
\item{sched}{Scheduler pointer}
}
\value{
Integer count of backends
}
\description{
Returns the number of backends managed by the scheduler.
}
