% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduler.R
\name{ggml_backend_sched_get_n_copies}
\alias{ggml_backend_sched_get_n_copies}
\title{Get number of tensor copies}
\usage{
ggml_backend_sched_get_n_copies(sched)
}
\arguments{
\item{sched}{Scheduler pointer}
}
\value{
Integer count of copies
}
\description{
Returns the number of tensor copies made in the last computed graph.
Copies occur when data needs to be transferred between backends.
}
