% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduler.R
\name{ggml_backend_sched_get_n_splits}
\alias{ggml_backend_sched_get_n_splits}
\title{Get number of graph splits}
\usage{
ggml_backend_sched_get_n_splits(sched)
}
\arguments{
\item{sched}{Scheduler pointer}
}
\value{
Integer count of splits
}
\description{
Returns the number of splits in the last computed graph.
Higher numbers indicate more distribution across backends.
}
