% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduler.R
\name{ggml_backend_sched_reset}
\alias{ggml_backend_sched_reset}
\title{Reset scheduler}
\usage{
ggml_backend_sched_reset(sched)
}
\arguments{
\item{sched}{Scheduler pointer}
}
\value{
NULL (invisible)
}
\description{
Resets the scheduler, deallocating all tensors.
Must be called before changing node backends or allocating a new graph.
}
