% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduler.R
\name{ggml_backend_sched_synchronize}
\alias{ggml_backend_sched_synchronize}
\title{Synchronize scheduler}
\usage{
ggml_backend_sched_synchronize(sched)
}
\arguments{
\item{sched}{Scheduler pointer}
}
\value{
NULL (invisible)
}
\description{
Waits for all asynchronous operations to complete.
}
