% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_functional.R
\name{ggml_batch_norm}
\alias{ggml_batch_norm}
\title{Create a Batch Normalization Layer Object}
\usage{
ggml_batch_norm(eps = 1e-05, name = NULL, trainable = TRUE)
}
\arguments{
\item{eps}{Small constant for numerical stability (default 1e-5).}

\item{name}{Optional character name.}

\item{trainable}{Logical.}
}
\value{
A \code{ggml_layer} object.
}
\description{
Create a Batch Normalization Layer Object
}
