% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{ggml_ceil_inplace}
\alias{ggml_ceil_inplace}
\title{Ceiling In-place (Graph)}
\usage{
ggml_ceil_inplace(ctx, a)
}
\arguments{
\item{ctx}{GGML context}

\item{a}{Input tensor (will be modified in-place)}
}
\value{
View of tensor a with ceiling values
}
\description{
Creates a graph node for in-place element-wise ceiling.
}
