% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{ggml_cont}
\alias{ggml_cont}
\title{Make Contiguous (Graph)}
\usage{
ggml_cont(ctx, a)
}
\arguments{
\item{ctx}{GGML context}

\item{a}{Input tensor}
}
\value{
Contiguous tensor
}
\description{
Makes a tensor contiguous in memory. Required after permute/transpose
before some operations.
}
