% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_functional.R
\name{ggml_conv_1d_layer}
\alias{ggml_conv_1d_layer}
\title{Create a Reusable Conv1D Layer Object}
\usage{
ggml_conv_1d_layer(
  filters,
  kernel_size,
  activation = NULL,
  strides = 1L,
  padding = "valid",
  name = NULL,
  trainable = TRUE
)
}
\arguments{
\item{filters}{Number of output filters.}

\item{kernel_size}{Integer kernel size.}

\item{activation}{Activation function name or NULL.}

\item{strides}{Integer stride (default 1).}

\item{padding}{\code{"valid"} or \code{"same"}.}

\item{name}{Optional character name.}

\item{trainable}{Logical.}
}
\value{
A \code{ggml_layer} object.
}
\description{
Builds a \emph{layer object} to be applied with \code{\link{ggml_apply}},
which is how a single set of convolution weights is shared between several
inputs.  This is the convolution counterpart of \code{\link{ggml_dense}}.
}
\details{
For the ordinary pipe-style form that appends a conv layer to a graph
(\code{x |> ggml_layer_conv_1d(...)}), see \code{\link{ggml_layer_conv_1d}}.
}
\examples{
\donttest{
# One convolution shared between two sequence inputs.
shared <- ggml_conv_1d_layer(filters = 4L, kernel_size = 3L)
x1 <- ggml_input(shape = c(16L, 2L))
x2 <- ggml_input(shape = c(16L, 2L))
o1 <- ggml_apply(x1, shared)
o2 <- ggml_apply(x2, shared)
stopifnot(identical(o1$layer_id, o2$layer_id))   # same weights
}
}
\seealso{
\code{\link{ggml_apply}}, \code{\link{ggml_dense}}
}
