% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{ggml_conv_transpose_2d_p0}
\alias{ggml_conv_transpose_2d_p0}
\title{Transposed 2D Convolution, zero padding (Graph)}
\usage{
ggml_conv_transpose_2d_p0(ctx, a, b, stride = 1L)
}
\arguments{
\item{ctx}{GGML context}

\item{a}{Convolution kernel tensor}

\item{b}{Input data tensor}

\item{stride}{Stride (default 1)}
}
\value{
Transposed convolved tensor
}
\description{
Applies transposed 2D convolution (deconvolution) with zero padding.
}
