% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggml.R
\name{ggml_cycles}
\alias{ggml_cycles}
\title{Get CPU Cycles}
\usage{
ggml_cycles()
}
\value{
Numeric value representing CPU cycles
}
\description{
Returns the current CPU cycle count. Useful for low-level benchmarking.
}
\examples{
\dontrun{
ggml_cycles()
}
}
