% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{ggml_diag_mask_inf_inplace}
\alias{ggml_diag_mask_inf_inplace}
\title{Diagonal Mask with -Inf In-place (Graph)}
\usage{
ggml_diag_mask_inf_inplace(ctx, a, n_past)
}
\arguments{
\item{ctx}{GGML context}

\item{a}{Input tensor (will be modified in-place)}

\item{n_past}{Number of past tokens}
}
\value{
View of input tensor with elements above diagonal set to -Inf
}
\description{
In-place version of ggml_diag_mask_inf. Returns a view of the input tensor.
}
