% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{ggml_dup_inplace}
\alias{ggml_dup_inplace}
\title{Duplicate Tensor In-place (Graph)}
\usage{
ggml_dup_inplace(ctx, a)
}
\arguments{
\item{ctx}{GGML context}

\item{a}{Input tensor}
}
\value{
View of tensor a
}
\description{
Creates a graph node for in-place tensor duplication.
Returns a view of the input tensor.
}
