% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{ggml_element_size}
\alias{ggml_element_size}
\title{Get Element Size}
\usage{
ggml_element_size(tensor)
}
\arguments{
\item{tensor}{Tensor pointer}
}
\value{
Element size in bytes
}
\description{
Returns the size of a single element in the tensor.
}
