% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{ggml_elu_inplace}
\alias{ggml_elu_inplace}
\title{ELU Activation In-place (Graph)}
\usage{
ggml_elu_inplace(ctx, a)
}
\arguments{
\item{ctx}{GGML context}

\item{a}{Input tensor (will be modified in-place)}
}
\value{
View of tensor a with ELU applied
}
\description{
Creates a graph node for in-place ELU (Exponential Linear Unit) activation.
}
