% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_functional.R
\name{ggml_embedding}
\alias{ggml_embedding}
\title{Create an Embedding Layer Object}
\usage{
ggml_embedding(vocab_size, dim, name = NULL, trainable = TRUE)
}
\arguments{
\item{vocab_size}{Number of distinct tokens.}

\item{dim}{Embedding dimension.}

\item{name}{Optional character name.}

\item{trainable}{Logical.}
}
\value{
A \code{ggml_layer} object.
}
\description{
Create an Embedding Layer Object
}
