% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory.R
\name{ggml_estimate_memory}
\alias{ggml_estimate_memory}
\title{Estimate Required Memory}
\usage{
ggml_estimate_memory(type = GGML_TYPE_F32, ne0, ne1 = 1, ne2 = 1, ne3 = 1)
}
\arguments{
\item{type}{Tensor type (GGML_TYPE_F32, etc)}

\item{ne0}{Size of dimension 0}

\item{ne1}{Size of dimension 1 (optional)}

\item{ne2}{Size of dimension 2 (optional)}

\item{ne3}{Size of dimension 3 (optional)}
}
\value{
Estimated memory in bytes
}
\description{
Helper function to estimate memory needed for a tensor
}
\examples{
\dontrun{
# For 1000x1000 F32 matrix
ggml_estimate_memory(GGML_TYPE_F32, 1000, 1000)
}
}
