% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{ggml_exp_inplace}
\alias{ggml_exp_inplace}
\title{Exponential In-place (Graph)}
\usage{
ggml_exp_inplace(ctx, a)
}
\arguments{
\item{ctx}{GGML context}

\item{a}{Input tensor (will be modified in-place)}
}
\value{
View of tensor a with exponential values
}
\description{
Creates a graph node for in-place element-wise exponential: e^x
}
