% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{ggml_flash_attn_ext_get_prec}
\alias{ggml_flash_attn_ext_get_prec}
\title{Get Flash-Attention Precision}
\usage{
ggml_flash_attn_ext_get_prec(tensor)
}
\arguments{
\item{tensor}{A flash-attention node}
}
\value{
The precision constant currently set on the node
}
\description{
Get Flash-Attention Precision
}
\seealso{
\code{\link{ggml_flash_attn_ext_set_prec}}
}
