% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{ggml_floor_inplace}
\alias{ggml_floor_inplace}
\title{Floor In-place (Graph)}
\usage{
ggml_floor_inplace(ctx, a)
}
\arguments{
\item{ctx}{GGML context}

\item{a}{Input tensor (will be modified in-place)}
}
\value{
View of tensor a with floor values
}
\description{
Creates a graph node for in-place element-wise floor.
}
