% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{ggml_ftype_to_ggml_type}
\alias{ggml_ftype_to_ggml_type}
\title{Convert ftype to ggml_type}
\usage{
ggml_ftype_to_ggml_type(ftype)
}
\arguments{
\item{ftype}{File type constant}
}
\value{
Integer GGML type
}
\description{
Converts a file type (ftype) to the corresponding GGML type.
Used when loading quantized models.
}
\seealso{
Other type_system: 
\code{\link{ggml_blck_size}()},
\code{\link{ggml_is_quantized}()},
\code{\link{ggml_type_name}()},
\code{\link{ggml_type_sizef}()}
}
\concept{type_system}
