% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{ggml_gallocr_free}
\alias{ggml_gallocr_free}
\title{Free Graph Allocator}
\usage{
ggml_gallocr_free(galloc)
}
\arguments{
\item{galloc}{Graph allocator object}
}
\description{
Frees a graph allocator and all associated buffers.
}
