% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{ggml_gelu_inplace}
\alias{ggml_gelu_inplace}
\title{GELU Activation In-place (Graph)}
\usage{
ggml_gelu_inplace(ctx, a)
}
\arguments{
\item{ctx}{GGML context}

\item{a}{Input tensor (will be modified in-place)}
}
\value{
View of tensor a with GELU applied
}
\description{
Creates a graph node for in-place GELU (Gaussian Error Linear Unit) activation.
CRITICAL for GPT models with memory efficiency.
}
