% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggml.R, R/tensors.R
\name{ggml_get_f32}
\alias{ggml_get_f32}
\title{Get F32 data}
\usage{
ggml_get_f32(tensor)

ggml_get_f32(tensor)
}
\arguments{
\item{tensor}{Tensor}
}
\value{
Numeric vector with tensor values

Numeric vector
}
\description{
Get F32 data

Get F32 Data
}
\examples{
\dontrun{
ctx <- ggml_init(1024 * 1024)
tensor <- ggml_new_tensor_1d(ctx, GGML_TYPE_F32, 5)
ggml_set_f32(tensor, c(1, 2, 3, 4, 5))
ggml_get_f32(tensor)
ggml_free(ctx)
}
\dontrun{
ctx <- ggml_init(1024 * 1024)
t <- ggml_new_tensor_1d(ctx, GGML_TYPE_F32, 5)
ggml_set_f32(t, c(1, 2, 3, 4, 5))
ggml_get_f32(t)
ggml_free(ctx)
}
}
