% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{ggml_get_f32_nd}
\alias{ggml_get_f32_nd}
\title{Get Single Float Value by N-D Index}
\usage{
ggml_get_f32_nd(tensor, i0, i1 = 0, i2 = 0, i3 = 0)
}
\arguments{
\item{tensor}{Tensor pointer}

\item{i0, i1, i2, i3}{Indices (0-based)}
}
\value{
Float value
}
\description{
Gets a single f32 value from the tensor at position [i0, i1, i2, i3].
Works with any tensor type (auto-converts to float).
}
