% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{ggml_get_first_tensor}
\alias{ggml_get_first_tensor}
\title{Get First Tensor from Context}
\usage{
ggml_get_first_tensor(ctx)
}
\arguments{
\item{ctx}{GGML context}
}
\value{
Tensor pointer or NULL
}
\description{
Get First Tensor from Context
}
