% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensors.R
\name{ggml_get_i32}
\alias{ggml_get_i32}
\title{Get I32 Data}
\usage{
ggml_get_i32(tensor)
}
\arguments{
\item{tensor}{Tensor of type GGML_TYPE_I32}
}
\value{
Integer vector
}
\description{
Gets integer data from an I32 tensor (e.g., from ggml_argmax)
}
\examples{
\dontrun{
ctx <- ggml_init(1024 * 1024)
pos <- ggml_new_tensor_1d(ctx, GGML_TYPE_I32, 10)
ggml_set_i32(pos, 0:9)
ggml_get_i32(pos)
ggml_free(ctx)
}
}
