% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threading.R
\name{ggml_get_n_threads}
\alias{ggml_get_n_threads}
\title{Get Number of Threads}
\usage{
ggml_get_n_threads()
}
\value{
Number of threads
}
\description{
Get the current number of threads for GGML operations
}
\examples{
\dontrun{
ggml_get_n_threads()
}
}
