% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{ggml_get_name}
\alias{ggml_get_name}
\title{Get Tensor Name}
\usage{
ggml_get_name(tensor)
}
\arguments{
\item{tensor}{Tensor pointer}
}
\value{
Character string name or NULL if not set
}
\description{
Retrieves the name of a tensor.
}
