% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{ggml_get_rel_pos}
\alias{ggml_get_rel_pos}
\title{Get Relative Position (Graph)}
\usage{
ggml_get_rel_pos(ctx, a, qh, kh)
}
\arguments{
\item{ctx}{GGML context}

\item{a}{Input tensor}

\item{qh}{Query height}

\item{kh}{Key height}
}
\value{
Relative-position tensor
}
\description{
Gathers relative-position rows for relative-position attention bias.
}
