% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{ggml_get_unary_op}
\alias{ggml_get_unary_op}
\title{Get Unary Operation from Tensor}
\usage{
ggml_get_unary_op(tensor)
}
\arguments{
\item{tensor}{Tensor pointer (must be a unary operation result)}
}
\value{
Integer unary operation type
}
\description{
Returns the unary operation type for a unary operation tensor.
}
\seealso{
Other op_info: 
\code{\link{ggml_op_desc}()},
\code{\link{ggml_op_name}()},
\code{\link{ggml_op_symbol}()},
\code{\link{ggml_unary_op_name}()}
}
\concept{op_info}
