% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\docType{data}
\name{GGML_GLU_OP_REGLU}
\alias{GGML_GLU_OP_REGLU}
\alias{GGML_GLU_OP_GEGLU}
\alias{GGML_GLU_OP_SWIGLU}
\alias{GGML_GLU_OP_SWIGLU_OAI}
\alias{GGML_GLU_OP_GEGLU_ERF}
\alias{GGML_GLU_OP_GEGLU_QUICK}
\title{GLU Operation Types}
\format{
Integer constants

An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.
}
\usage{
GGML_GLU_OP_REGLU

GGML_GLU_OP_GEGLU

GGML_GLU_OP_SWIGLU

GGML_GLU_OP_SWIGLU_OAI

GGML_GLU_OP_GEGLU_ERF

GGML_GLU_OP_GEGLU_QUICK
}
\description{
Constants for GLU (Gated Linear Unit) operation types.
Used with ggml_glu() and ggml_glu_split().
}
\details{
\itemize{
  \item \code{GGML_GLU_OP_REGLU}: ReGLU - ReLU gating
  \item \code{GGML_GLU_OP_GEGLU}: GeGLU - GELU gating (used in GPT-NeoX, etc.)
  \item \code{GGML_GLU_OP_SWIGLU}: SwiGLU - SiLU/Swish gating (used in LLaMA)
  \item \code{GGML_GLU_OP_GEGLU_QUICK}: GeGLU with fast approximation
}
}
\keyword{datasets}
