% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{ggml_graph_n_nodes}
\alias{ggml_graph_n_nodes}
\title{Get Number of Nodes in Graph}
\usage{
ggml_graph_n_nodes(graph)
}
\arguments{
\item{graph}{Graph object}
}
\value{
Integer number of nodes
}
\description{
Returns the number of computation nodes in the graph
}
