% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{ggml_graph_overhead}
\alias{ggml_graph_overhead}
\title{Get Graph Overhead}
\usage{
ggml_graph_overhead()
}
\value{
Size in bytes
}
\description{
Returns the memory overhead required for a computation graph
}
