% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{ggml_im2col}
\alias{ggml_im2col}
\title{Image to Column (Graph)}
\usage{
ggml_im2col(
  ctx,
  a,
  b,
  s0,
  s1,
  p0,
  p1,
  d0,
  d1,
  is_2D = TRUE,
  dst_type = GGML_TYPE_F16
)
}
\arguments{
\item{ctx}{GGML context}

\item{a}{Convolution kernel tensor}

\item{b}{Input data tensor}

\item{s0}{Stride dimension 0}

\item{s1}{Stride dimension 1}

\item{p0}{Padding dimension 0}

\item{p1}{Padding dimension 1}

\item{d0}{Dilation dimension 0}

\item{d1}{Dilation dimension 1}

\item{is_2D}{Whether this is a 2D operation (default TRUE)}

\item{dst_type}{Output type (default GGML_TYPE_F16)}
}
\value{
Transformed tensor in column format
}
\description{
Transforms image data into column format for efficient convolution.
This is a low-level operation used internally by convolution implementations.
}
