% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggml.R
\name{ggml_is_available}
\alias{ggml_is_available}
\title{Check if GGML is available}
\usage{
ggml_is_available()
}
\value{
TRUE if GGML library is loaded
}
\description{
Check if GGML is available
}
\examples{
\dontrun{
ggml_is_available()
}
}
