% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{ggml_is_contiguous_0}
\alias{ggml_is_contiguous_0}
\title{Check Tensor Contiguity (Dimension 0)}
\usage{
ggml_is_contiguous_0(tensor)
}
\arguments{
\item{tensor}{Tensor pointer}
}
\value{
Logical indicating contiguity
}
\description{
Check if tensor is contiguous. Same as \code{ggml_is_contiguous}.
}
\seealso{
Other tensor_layout: 
\code{\link{ggml_are_same_stride}()},
\code{\link{ggml_can_repeat}()},
\code{\link{ggml_count_equal}()},
\code{\link{ggml_is_contiguous_1}()},
\code{\link{ggml_is_contiguous_2}()},
\code{\link{ggml_is_contiguous_channels}()},
\code{\link{ggml_is_contiguous_rows}()},
\code{\link{ggml_is_contiguously_allocated}()}
}
\concept{tensor_layout}
