% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensors.R
\name{ggml_is_transposed}
\alias{ggml_is_transposed}
\title{Check if Tensor is Transposed}
\usage{
ggml_is_transposed(tensor)
}
\arguments{
\item{tensor}{Tensor pointer}
}
\value{
Logical
}
\description{
Returns TRUE if tensor has been transposed
}
\examples{
\dontrun{
ctx <- ggml_init(1024 * 1024)
t <- ggml_new_tensor_2d(ctx, GGML_TYPE_F32, 10, 20)
ggml_is_transposed(t)
ggml_free(ctx)
}
}
