% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{ggml_log_inplace}
\alias{ggml_log_inplace}
\title{Natural Logarithm In-place (Graph)}
\usage{
ggml_log_inplace(ctx, a)
}
\arguments{
\item{ctx}{GGML context}

\item{a}{Input tensor (will be modified in-place)}
}
\value{
View of tensor a with log values
}
\description{
Creates a graph node for in-place element-wise natural logarithm.
}
