% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{ggml_log_set_default}
\alias{ggml_log_set_default}
\title{Restore Default GGML Logging}
\usage{
ggml_log_set_default()
}
\value{
NULL invisibly
}
\description{
Restores GGML to default logging behavior (stderr output).
}
\seealso{
Other logging: 
\code{\link{ggml_abort_is_r_enabled}()},
\code{\link{ggml_log_is_r_enabled}()},
\code{\link{ggml_log_set_r}()},
\code{\link{ggml_set_abort_callback_default}()},
\code{\link{ggml_set_abort_callback_r}()}
}
\concept{logging}
