% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{ggml_log_set_r}
\alias{ggml_log_set_r}
\title{Enable R-compatible GGML Logging}
\usage{
ggml_log_set_r()
}
\value{
NULL invisibly
}
\description{
Redirects GGML log messages to R's message system:
- INFO/DEBUG messages go to stdout (via Rprintf)
- WARN/ERROR messages go to stderr (via REprintf)
}
\examples{
\donttest{
ggml_log_set_r()
# Now GGML messages will appear in R console
}
}
\seealso{
Other logging: 
\code{\link{ggml_abort_is_r_enabled}()},
\code{\link{ggml_log_is_r_enabled}()},
\code{\link{ggml_log_set_default}()},
\code{\link{ggml_set_abort_callback_default}()},
\code{\link{ggml_set_abort_callback_r}()}
}
\concept{logging}
