% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_functional.R
\name{ggml_lstm}
\alias{ggml_lstm}
\title{Create an LSTM Layer Object}
\usage{
ggml_lstm(
  units,
  return_sequences = FALSE,
  activation = "tanh",
  recurrent_activation = "sigmoid",
  name = NULL,
  trainable = TRUE
)
}
\arguments{
\item{units}{Integer, number of hidden units.}

\item{return_sequences}{Logical.}

\item{activation}{Cell gate activation (default \code{"tanh"}).}

\item{recurrent_activation}{Recurrent gate activation (default \code{"sigmoid"}).}

\item{name}{Optional character name.}

\item{trainable}{Logical.}
}
\value{
A \code{ggml_layer} object.
}
\description{
Create an LSTM Layer Object
}
