% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_model.R
\name{ggml_model_sequential}
\alias{ggml_model_sequential}
\title{Create a Sequential Neural Network Model}
\usage{
ggml_model_sequential()
}
\value{
A ggml_sequential_model object
}
\description{
Creates an empty sequential model that layers can be added to using
pipe (\code{|>}) operators.
}
\examples{
\dontrun{
model <- ggml_model_sequential() |>
  ggml_layer_conv_2d(32, c(3,3), activation = "relu",
                     input_shape = c(28, 28, 1)) |>
  ggml_layer_max_pooling_2d(c(2,2)) |>
  ggml_layer_flatten() |>
  ggml_layer_dense(128, activation = "relu") |>
  ggml_layer_dense(10, activation = "softmax")
}
}
