% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensors.R
\name{ggml_n_dims}
\alias{ggml_n_dims}
\title{Get Number of Dimensions}
\usage{
ggml_n_dims(tensor)
}
\arguments{
\item{tensor}{Tensor pointer}
}
\value{
Integer number of dimensions (1-4)
}
\description{
Returns the number of dimensions of a tensor
}
\examples{
\dontrun{
ctx <- ggml_init(1024 * 1024)
t <- ggml_new_tensor_2d(ctx, GGML_TYPE_F32, 10, 20)
ggml_n_dims(t)
ggml_free(ctx)
}
}
