% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_ops.R
\docType{data}
\name{GGML_N_TASKS_MAX}
\alias{GGML_N_TASKS_MAX}
\title{Maximum Task Count for Custom Ops}
\format{
An object of class \code{integer} of length 1.
}
\usage{
GGML_N_TASKS_MAX
}
\description{
Sentinel for the \code{n_tasks} argument of \code{\link{ggml_custom}} and
\code{\link{ggml_custom_inplace}}, meaning "use as many threads as the
backend allows" (mirrors \code{GGML_N_TASKS_MAX} in ggml.h).
}
\keyword{datasets}
