% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggml.R, R/tensors.R
\name{ggml_nbytes}
\alias{ggml_nbytes}
\title{Get number of bytes}
\usage{
ggml_nbytes(tensor)

ggml_nbytes(tensor)
}
\arguments{
\item{tensor}{Tensor}
}
\value{
Integer number of bytes

Integer number of bytes
}
\description{
Get number of bytes

Get Number of Bytes
}
\examples{
\dontrun{
ctx <- ggml_init(1024 * 1024)
tensor <- ggml_new_tensor_1d(ctx, GGML_TYPE_F32, 10)
ggml_nbytes(tensor)
ggml_free(ctx)
}
\dontrun{
ctx <- ggml_init(1024 * 1024)
t <- ggml_new_tensor_1d(ctx, GGML_TYPE_F32, 10)
ggml_nbytes(t)
ggml_free(ctx)
}
}
