% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{ggml_neg}
\alias{ggml_neg}
\title{Negation (Graph)}
\usage{
ggml_neg(ctx, a)
}
\arguments{
\item{ctx}{GGML context}

\item{a}{Input tensor}
}
\value{
Tensor representing the negation operation
}
\description{
Creates a graph node for element-wise negation: -x
}
