% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{ggml_neg_inplace}
\alias{ggml_neg_inplace}
\title{Negation In-place (Graph)}
\usage{
ggml_neg_inplace(ctx, a)
}
\arguments{
\item{ctx}{GGML context}

\item{a}{Input tensor (will be modified in-place)}
}
\value{
View of tensor a with negated values
}
\description{
Creates a graph node for in-place element-wise negation: -x
}
