% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggml.R, R/tensors.R
\name{ggml_nelements}
\alias{ggml_nelements}
\title{Get number of elements}
\usage{
ggml_nelements(tensor)

ggml_nelements(tensor)
}
\arguments{
\item{tensor}{Tensor}
}
\value{
Integer number of elements

Integer number of elements
}
\description{
Get number of elements

Get Number of Elements
}
\examples{
\donttest{
ctx <- ggml_init(1024 * 1024)
tensor <- ggml_new_tensor_2d(ctx, GGML_TYPE_F32, 10, 20)
ggml_nelements(tensor)
ggml_free(ctx)
}
\donttest{
ctx <- ggml_init(1024 * 1024)
t <- ggml_new_tensor_2d(ctx, GGML_TYPE_F32, 10, 20)
ggml_nelements(t)
ggml_free(ctx)
}
}
