% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggml.R, R/tensors.R
\name{ggml_new_tensor_1d}
\alias{ggml_new_tensor_1d}
\title{Create 1D tensor}
\usage{
ggml_new_tensor_1d(ctx, type = GGML_TYPE_F32, ne0)

ggml_new_tensor_1d(ctx, type = GGML_TYPE_F32, ne0)
}
\arguments{
\item{ctx}{GGML context}

\item{type}{Data type}

\item{ne0}{Size}
}
\value{
Tensor pointer

Tensor pointer
}
\description{
Create 1D tensor

Create 1D Tensor
}
\examples{
\donttest{
ctx <- ggml_init(1024 * 1024)
tensor <- ggml_new_tensor_1d(ctx, GGML_TYPE_F32, 10)
ggml_nelements(tensor)
ggml_free(ctx)
}
}
