% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggml.R, R/tensors.R
\name{ggml_new_tensor_2d}
\alias{ggml_new_tensor_2d}
\title{Create 2D tensor}
\usage{
ggml_new_tensor_2d(ctx, type = GGML_TYPE_F32, ne0, ne1)

ggml_new_tensor_2d(ctx, type = GGML_TYPE_F32, ne0, ne1)
}
\arguments{
\item{ctx}{GGML context}

\item{type}{Data type}

\item{ne0}{Rows}

\item{ne1}{Columns}
}
\value{
Tensor pointer

Tensor pointer
}
\description{
Create 2D tensor

Create 2D Tensor
}
\examples{
\donttest{
ctx <- ggml_init(1024 * 1024)
tensor <- ggml_new_tensor_2d(ctx, GGML_TYPE_F32, 10, 20)
ggml_nelements(tensor)
ggml_free(ctx)
}
}
