% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{ggml_nrows}
\alias{ggml_nrows}
\title{Get Number of Rows}
\usage{
ggml_nrows(tensor)
}
\arguments{
\item{tensor}{Tensor pointer}
}
\value{
Number of rows
}
\description{
Returns the number of rows in a tensor (product of all dimensions except ne[0]).
}
