% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{ggml_op_desc}
\alias{ggml_op_desc}
\title{Get Operation Description from Tensor}
\usage{
ggml_op_desc(tensor)
}
\arguments{
\item{tensor}{Tensor pointer}
}
\value{
Character string describing the operation
}
\description{
Returns a description of the operation that produces a tensor.
}
\seealso{
Other op_info: 
\code{\link{ggml_get_unary_op}()},
\code{\link{ggml_op_name}()},
\code{\link{ggml_op_symbol}()},
\code{\link{ggml_unary_op_name}()}
}
\concept{op_info}
