% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{ggml_pad_reflect_1d}
\alias{ggml_pad_reflect_1d}
\title{Reflective 1D Padding (Graph)}
\usage{
ggml_pad_reflect_1d(ctx, a, p0, p1)
}
\arguments{
\item{ctx}{GGML context}

\item{a}{Input tensor}

\item{p0}{Left padding}

\item{p1}{Right padding}
}
\value{
Padded tensor
}
\description{
Pads the first dimension of a tensor using reflection of its values.
}
