% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{ggml_pool_2d}
\alias{ggml_pool_2d}
\title{2D Pooling (Graph)}
\usage{
ggml_pool_2d(ctx, a, op, k0, k1, s0 = k0, s1 = k1, p0 = 0, p1 = 0)
}
\arguments{
\item{ctx}{GGML context}

\item{a}{Input tensor}

\item{op}{Pool operation: GGML_OP_POOL_MAX (0) or GGML_OP_POOL_AVG (1)}

\item{k0}{Kernel size dimension 0}

\item{k1}{Kernel size dimension 1}

\item{s0}{Stride dimension 0 (default = k0)}

\item{s1}{Stride dimension 1 (default = k1)}

\item{p0}{Padding dimension 0 (default 0)}

\item{p1}{Padding dimension 1 (default 0)}
}
\value{
Pooled tensor
}
\description{
Applies 2D pooling operation.
}
