% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory.R
\name{ggml_print_objects}
\alias{ggml_print_objects}
\title{Print Objects in Context}
\usage{
ggml_print_objects(ctx)
}
\arguments{
\item{ctx}{GGML context}
}
\value{
NULL (invisible)
}
\description{
Debug function to print all objects (tensors) in the context
}
\examples{
\dontrun{
ctx <- ggml_init(1024 * 1024)
t <- ggml_new_tensor_1d(ctx, GGML_TYPE_F32, 10)
ggml_print_objects(ctx)
ggml_free(ctx)
}
}
