% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{ggml_quantize_free}
\alias{ggml_quantize_free}
\title{Free Quantization Resources}
\usage{
ggml_quantize_free()
}
\value{
NULL invisibly
}
\description{
Frees any memory allocated by quantization.
Call at end of program to avoid memory leaks.
}
