% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{ggml_quantize_init}
\alias{ggml_quantize_init}
\title{Initialize Quantization Tables}
\usage{
ggml_quantize_init(type)
}
\arguments{
\item{type}{GGML type (e.g., GGML_TYPE_Q4_0)}
}
\value{
NULL invisibly
}
\description{
Initializes quantization tables for a given type.
Called automatically by ggml_quantize_chunk, but can be called manually.
}
