% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{ggml_relu_inplace}
\alias{ggml_relu_inplace}
\title{ReLU Activation In-place (Graph)}
\usage{
ggml_relu_inplace(ctx, a)
}
\arguments{
\item{ctx}{GGML context}

\item{a}{Input tensor (will be modified in-place)}
}
\value{
View of tensor a with ReLU applied
}
\description{
Creates a graph node for in-place ReLU activation: max(0, x)
}
