% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{ggml_rope_inplace}
\alias{ggml_rope_inplace}
\title{Rotary Position Embedding In-place (Graph)}
\usage{
ggml_rope_inplace(ctx, a, b, n_dims, mode = 0L)
}
\arguments{
\item{ctx}{GGML context}

\item{a}{Input tensor (will be modified in-place)}

\item{b}{Position tensor (int32)}

\item{n_dims}{Number of dimensions to apply rotation to}

\item{mode}{RoPE mode}
}
\value{
View of input tensor with RoPE applied
}
\description{
In-place version of ggml_rope. Returns a view of the input tensor.
}
