% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{ggml_round}
\alias{ggml_round}
\title{Round (Graph)}
\usage{
ggml_round(ctx, a)
}
\arguments{
\item{ctx}{GGML context}

\item{a}{Input tensor}
}
\value{
Tensor representing the round operation
}
\description{
Creates a graph node for element-wise rounding: round(x)
}
