% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{ggml_round_inplace}
\alias{ggml_round_inplace}
\title{Round In-place (Graph)}
\usage{
ggml_round_inplace(ctx, a)
}
\arguments{
\item{ctx}{GGML context}

\item{a}{Input tensor (will be modified in-place)}
}
\value{
View of tensor a with rounded values
}
\description{
Creates a graph node for in-place element-wise rounding.
}
