% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{ggml_scale_inplace}
\alias{ggml_scale_inplace}
\title{Scale Tensor In-place (Graph)}
\usage{
ggml_scale_inplace(ctx, a, s)
}
\arguments{
\item{ctx}{GGML context}

\item{a}{Input tensor (will be modified in-place)}

\item{s}{Scalar value to multiply by}
}
\value{
View of tensor a with scaled values
}
\description{
Creates a graph node for in-place scaling: a * s
}
