% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{ggml_schedule_cosine_decay}
\alias{ggml_schedule_cosine_decay}
\title{Cosine annealing LR scheduler}
\usage{
ggml_schedule_cosine_decay(eta_min = 0, T_max = NULL)
}
\arguments{
\item{eta_min}{Minimum LR at end of schedule}

\item{T_max}{Total number of epochs (defaults to nepoch from fit state)}
}
\value{
List with on_epoch_begin function
}
\description{
Anneals LR from initial value to `eta_min` following a cosine curve.
}
\seealso{
Other callbacks: 
\code{\link{ggml_callback_early_stopping}()},
\code{\link{ggml_schedule_reduce_on_plateau}()},
\code{\link{ggml_schedule_step_decay}()}
}
\concept{callbacks}
