% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{ggml_schedule_step_decay}
\alias{ggml_schedule_step_decay}
\title{Step decay LR scheduler}
\usage{
ggml_schedule_step_decay(step_size = 10, gamma = 0.1)
}
\arguments{
\item{step_size}{Reduce LR every this many epochs}

\item{gamma}{Multiplicative factor of LR reduction}
}
\value{
List with on_epoch_begin function
}
\description{
Reduces LR by a factor every `step_size` epochs.
}
\seealso{
Other callbacks: 
\code{\link{ggml_callback_early_stopping}()},
\code{\link{ggml_schedule_cosine_decay}()},
\code{\link{ggml_schedule_reduce_on_plateau}()}
}
\concept{callbacks}
