% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{ggml_set}
\alias{ggml_set}
\title{Set Tensor Region (Graph)}
\usage{
ggml_set(ctx, a, b, nb1, nb2, nb3, offset)
}
\arguments{
\item{ctx}{GGML context}

\item{a}{Destination tensor}

\item{b}{Source tensor (data to copy)}

\item{nb1}{Stride for dimension 1 (in bytes)}

\item{nb2}{Stride for dimension 2 (in bytes)}

\item{nb3}{Stride for dimension 3 (in bytes)}

\item{offset}{Byte offset in destination tensor}
}
\value{
Tensor representing the set operation
}
\description{
Copies tensor b into tensor a at a specified offset.
This allows writing to a portion of a tensor.
}
