% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{ggml_set_abort_callback_default}
\alias{ggml_set_abort_callback_default}
\title{Restore Default Abort Behavior}
\usage{
ggml_set_abort_callback_default()
}
\value{
NULL invisibly
}
\description{
Restores GGML to default abort behavior (prints to stderr and aborts).
}
\seealso{
Other logging: 
\code{\link{ggml_abort_is_r_enabled}()},
\code{\link{ggml_log_is_r_enabled}()},
\code{\link{ggml_log_set_default}()},
\code{\link{ggml_log_set_r}()},
\code{\link{ggml_set_abort_callback_r}()}
}
\concept{logging}
