% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggml.R, R/tensors.R
\name{ggml_set_f32}
\alias{ggml_set_f32}
\title{Set F32 data}
\usage{
ggml_set_f32(tensor, data)

ggml_set_f32(tensor, data)
}
\arguments{
\item{tensor}{Tensor}

\item{data}{Numeric vector}
}
\value{
NULL (invisible)

NULL (invisible)
}
\description{
Set F32 data

Set F32 Data
}
\examples{
\donttest{
ctx <- ggml_init(1024 * 1024)
tensor <- ggml_new_tensor_1d(ctx, GGML_TYPE_F32, 5)
ggml_set_f32(tensor, c(1, 2, 3, 4, 5))
ggml_get_f32(tensor)
ggml_free(ctx)
}
\donttest{
ctx <- ggml_init(1024 * 1024)
t <- ggml_new_tensor_1d(ctx, GGML_TYPE_F32, 5)
ggml_set_f32(t, c(1, 2, 3, 4, 5))
ggml_get_f32(t)
ggml_free(ctx)
}
}
