% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensors.R
\name{ggml_set_i32}
\alias{ggml_set_i32}
\title{Set I32 Data}
\usage{
ggml_set_i32(tensor, data)
}
\arguments{
\item{tensor}{Tensor of type GGML_TYPE_I32}

\item{data}{Integer vector}
}
\value{
NULL (invisible)
}
\description{
Sets integer data in an I32 tensor. Used for indices (ggml_get_rows)
and position tensors (ggml_rope).
}
\examples{
\dontrun{
ctx <- ggml_init(1024 * 1024)
pos <- ggml_new_tensor_1d(ctx, GGML_TYPE_I32, 10)
ggml_set_i32(pos, 0:9)
ggml_get_i32(pos)
ggml_free(ctx)
}
}
