% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{ggml_set_input}
\alias{ggml_set_input}
\title{Mark Tensor as Input}
\usage{
ggml_set_input(tensor)
}
\arguments{
\item{tensor}{Tensor pointer}
}
\value{
The tensor (for chaining)
}
\description{
Mark Tensor as Input
}
