% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threading.R
\name{ggml_set_n_threads}
\alias{ggml_set_n_threads}
\title{Set Number of Threads}
\usage{
ggml_set_n_threads(n_threads)
}
\arguments{
\item{n_threads}{Number of threads to use}
}
\value{
Number of threads set
}
\description{
Set the number of threads for GGML operations
}
\examples{
\dontrun{
# Use 4 threads
ggml_set_n_threads(4)

# Use all available cores
ggml_set_n_threads(parallel::detectCores())
}
}
