% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{ggml_set_op_params}
\alias{ggml_set_op_params}
\title{Set Tensor Operation Parameters}
\usage{
ggml_set_op_params(tensor, params)
}
\arguments{
\item{tensor}{External pointer to tensor}

\item{params}{Raw vector of parameters (max 64 bytes)}
}
\value{
NULL invisibly
}
\description{
Sets the raw op_params bytes for a tensor.
}
\seealso{
Other tensor: 
\code{\link{ggml_are_same_layout}()},
\code{\link{ggml_get_op_params}()},
\code{\link{ggml_get_op_params_f32}()},
\code{\link{ggml_get_op_params_i32}()},
\code{\link{ggml_set_op_params_f32}()},
\code{\link{ggml_set_op_params_i32}()}
}
\concept{tensor}
