% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{ggml_set_op_params_i32}
\alias{ggml_set_op_params_i32}
\title{Set Integer Op Parameter}
\usage{
ggml_set_op_params_i32(tensor, index, value)
}
\arguments{
\item{tensor}{External pointer to tensor}

\item{index}{0-based index (0-15 for 64-byte op_params)}

\item{value}{Integer value to set}
}
\value{
NULL invisibly
}
\description{
Sets a single int32 value in tensor op_params at given index.
}
\seealso{
Other tensor: 
\code{\link{ggml_are_same_layout}()},
\code{\link{ggml_get_op_params}()},
\code{\link{ggml_get_op_params_f32}()},
\code{\link{ggml_get_op_params_i32}()},
\code{\link{ggml_set_op_params}()},
\code{\link{ggml_set_op_params_f32}()}
}
\concept{tensor}
