% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{ggml_set_output}
\alias{ggml_set_output}
\title{Mark Tensor as Output}
\usage{
ggml_set_output(tensor)
}
\arguments{
\item{tensor}{Tensor pointer}
}
\value{
The tensor (for chaining)
}
\description{
Mark Tensor as Output
}
