% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{ggml_sigmoid_inplace}
\alias{ggml_sigmoid_inplace}
\title{Sigmoid Activation In-place (Graph)}
\usage{
ggml_sigmoid_inplace(ctx, a)
}
\arguments{
\item{ctx}{GGML context}

\item{a}{Input tensor (will be modified in-place)}
}
\value{
View of tensor a with sigmoid applied
}
\description{
Creates a graph node for in-place sigmoid activation: 1 / (1 + e^(-x))
}
