% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{ggml_soft_max_add_sinks}
\alias{ggml_soft_max_add_sinks}
\title{Attach Attention Sinks to a Softmax Node}
\usage{
ggml_soft_max_add_sinks(tensor, sinks)
}
\arguments{
\item{tensor}{A softmax node}

\item{sinks}{Sink logits, one per head}
}
\value{
The tensor (for chaining)
}
\description{
The \code{\link{ggml_soft_max_ext}} counterpart of
\code{\link{ggml_flash_attn_ext_add_sinks}}.
}
