% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{ggml_softplus_inplace}
\alias{ggml_softplus_inplace}
\title{Softplus Activation In-place (Graph)}
\usage{
ggml_softplus_inplace(ctx, a)
}
\arguments{
\item{ctx}{GGML context}

\item{a}{Input tensor (will be modified in-place)}
}
\value{
View of tensor a with softplus applied
}
\description{
Creates a graph node for in-place softplus activation: log(1 + e^x)
}
