% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\docType{data}
\name{GGML_SORT_ORDER_ASC}
\alias{GGML_SORT_ORDER_ASC}
\alias{GGML_SORT_ORDER_DESC}
\title{Sort Order Constants}
\format{
Integer constants

An object of class \code{integer} of length 1.
}
\usage{
GGML_SORT_ORDER_ASC

GGML_SORT_ORDER_DESC
}
\description{
Constants for specifying sort order in argsort operations.
}
\examples{
\dontrun{
GGML_SORT_ORDER_ASC   # 0 - Ascending order
GGML_SORT_ORDER_DESC  # 1 - Descending order
}
}
\keyword{datasets}
