% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{ggml_sqr}
\alias{ggml_sqr}
\title{Square (Graph)}
\usage{
ggml_sqr(ctx, a)
}
\arguments{
\item{ctx}{GGML context}

\item{a}{Input tensor}
}
\value{
Tensor representing the square operation
}
\description{
Creates a graph node for element-wise squaring: x^2
}
