% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{ggml_sqr_inplace}
\alias{ggml_sqr_inplace}
\title{Square In-place (Graph)}
\usage{
ggml_sqr_inplace(ctx, a)
}
\arguments{
\item{ctx}{GGML context}

\item{a}{Input tensor (will be modified in-place)}
}
\value{
View of tensor a with squared values
}
\description{
Creates a graph node for in-place element-wise square: x^2
}
