% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{ggml_sqrt_inplace}
\alias{ggml_sqrt_inplace}
\title{Square Root In-place (Graph)}
\usage{
ggml_sqrt_inplace(ctx, a)
}
\arguments{
\item{ctx}{GGML context}

\item{a}{Input tensor (will be modified in-place)}
}
\value{
View of tensor a with square root values
}
\description{
Creates a graph node for in-place element-wise square root.
}
