% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{ggml_tanh_inplace}
\alias{ggml_tanh_inplace}
\title{Tanh Activation In-place (Graph)}
\usage{
ggml_tanh_inplace(ctx, a)
}
\arguments{
\item{ctx}{GGML context}

\item{a}{Input tensor (will be modified in-place)}
}
\value{
View of tensor a with tanh applied
}
\description{
Creates a graph node for in-place hyperbolic tangent activation.
}
