% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{ggml_tensor_copy}
\alias{ggml_tensor_copy}
\title{Copy Tensor Data}
\usage{
ggml_tensor_copy(dst, src)
}
\arguments{
\item{dst}{Destination tensor}

\item{src}{Source tensor}
}
\value{
NULL (invisible)
}
\description{
Copies raw data from src tensor to dst tensor (must be same size).
}
