% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{ggml_tensor_nb}
\alias{ggml_tensor_nb}
\title{Get Tensor Strides (nb)}
\usage{
ggml_tensor_nb(tensor)
}
\arguments{
\item{tensor}{Tensor pointer}
}
\value{
Numeric vector of 4 stride values (nb0, nb1, nb2, nb3)
}
\description{
Returns the byte strides for each dimension of the tensor.
}
