% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{ggml_tensor_num}
\alias{ggml_tensor_num}
\title{Count Tensors in Context}
\usage{
ggml_tensor_num(ctx)
}
\arguments{
\item{ctx}{GGML context}
}
\value{
Number of tensors
}
\description{
Counts the number of tensors allocated in a context.
}
