% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory.R
\name{ggml_tensor_overhead}
\alias{ggml_tensor_overhead}
\title{Get Tensor Overhead}
\usage{
ggml_tensor_overhead()
}
\value{
Size in bytes
}
\description{
Returns the memory overhead (metadata) for each tensor in bytes
}
\examples{
\donttest{
ggml_tensor_overhead()
}
}
