% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensors.R
\name{ggml_tensor_type}
\alias{ggml_tensor_type}
\title{Get Tensor Type}
\usage{
ggml_tensor_type(tensor)
}
\arguments{
\item{tensor}{Tensor pointer}
}
\value{
Integer type code (0 = F32, 1 = F16, etc.)
}
\description{
Returns the data type of a tensor as an integer code
}
\examples{
\donttest{
ctx <- ggml_init(1024 * 1024)
t <- ggml_new_tensor_1d(ctx, GGML_TYPE_F32, 10)
ggml_tensor_type(t)
ggml_free(ctx)
}
}
