% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{ggml_timestep_embedding}
\alias{ggml_timestep_embedding}
\title{Timestep Embedding (Graph Operation)}
\usage{
ggml_timestep_embedding(ctx, timesteps, dim, max_period = 10000L)
}
\arguments{
\item{ctx}{GGML context}

\item{timesteps}{Input tensor of timestep values [N]}

\item{dim}{Embedding dimension}

\item{max_period}{Maximum period for sinusoidal embedding (default 10000)}
}
\value{
Tensor of shape [N, dim] with timestep embeddings
}
\description{
Creates sinusoidal timestep embeddings as used in diffusion models.
Reference: CompVis/stable-diffusion util.py timestep_embedding
}
