% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{ggml_type_name}
\alias{ggml_type_name}
\title{Get Type Name}
\usage{
ggml_type_name(type)
}
\arguments{
\item{type}{GGML type constant (e.g., GGML_TYPE_F32)}
}
\value{
Character string with type name
}
\description{
Returns the string name of a GGML type.
}
\examples{
ggml_type_name(GGML_TYPE_F32)  # "f32"
ggml_type_name(GGML_TYPE_Q4_0) # "q4_0"
}
\seealso{
Other type_system: 
\code{\link{ggml_blck_size}()},
\code{\link{ggml_ftype_to_ggml_type}()},
\code{\link{ggml_is_quantized}()},
\code{\link{ggml_type_sizef}()}
}
\concept{type_system}
