% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{ggml_unary_op_name}
\alias{ggml_unary_op_name}
\title{Get Unary Operation Name}
\usage{
ggml_unary_op_name(op)
}
\arguments{
\item{op}{GGML unary operation constant}
}
\value{
Character string with operation name
}
\description{
Returns the string name of a GGML unary operation.
}
\seealso{
Other op_info: 
\code{\link{ggml_get_unary_op}()},
\code{\link{ggml_op_desc}()},
\code{\link{ggml_op_name}()},
\code{\link{ggml_op_symbol}()}
}
\concept{op_info}
