% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory.R
\name{ggml_used_mem}
\alias{ggml_used_mem}
\title{Get Used Memory}
\usage{
ggml_used_mem(ctx)
}
\arguments{
\item{ctx}{GGML context}
}
\value{
Used memory in bytes
}
\description{
Returns the amount of memory currently used in the context
}
\examples{
\donttest{
ctx <- ggml_init(1024 * 1024)
a <- ggml_new_tensor_1d(ctx, GGML_TYPE_F32, 100)
ggml_used_mem(ctx)
ggml_free(ctx)
}
}
