% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vulkan.R
\name{ggml_vulkan_device_count}
\alias{ggml_vulkan_device_count}
\title{Get number of Vulkan devices}
\usage{
ggml_vulkan_device_count()
}
\value{
Integer count of Vulkan devices (0 if Vulkan not available)
}
\description{
Returns the number of available Vulkan-capable GPU devices.
}
\examples{
\donttest{
if (ggml_vulkan_available()) {
  ggml_vulkan_device_count()
}
}
}
