% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vulkan.R
\name{ggml_vulkan_list_devices}
\alias{ggml_vulkan_list_devices}
\title{List all Vulkan devices}
\usage{
ggml_vulkan_list_devices()
}
\value{
List of device information (index, name, memory)
}
\description{
Returns detailed information about all available Vulkan devices.
}
\examples{
\dontrun{
if (ggml_vulkan_available() && ggml_vulkan_device_count() > 0) {
  devices <- ggml_vulkan_list_devices()
  print(devices)
}
}
}
