% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{ggml_win_part}
\alias{ggml_win_part}
\title{Window Partition (Graph)}
\usage{
ggml_win_part(ctx, a, w)
}
\arguments{
\item{ctx}{GGML context}

\item{a}{Input tensor}

\item{w}{Window size}
}
\value{
Partitioned tensor
}
\description{
Partitions a tensor into non-overlapping windows of size \code{w}.
}
