% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gguf.R
\name{gguf_tensor_names}
\alias{gguf_tensor_names}
\title{List Tensor Names in a GGUF File}
\usage{
gguf_tensor_names(x)
}
\arguments{
\item{x}{A \code{gguf} object.}
}
\value{
Character vector of tensor names.
}
\description{
List Tensor Names in a GGUF File
}
