% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autograd.R
\name{is_ag_tensor}
\alias{is_ag_tensor}
\title{Check if object is an ag_tensor}
\usage{
is_ag_tensor(x)
}
\arguments{
\item{x}{Any R object to test.}
}
\value{
\code{TRUE} if \code{x} is an \code{ag_tensor} object, \code{FALSE} otherwise.
}
\description{
Check if object is an ag_tensor
}
\keyword{internal}
