% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_layers.R
\name{nn_build_embedding}
\alias{nn_build_embedding}
\title{Build embedding forward pass}
\usage{
nn_build_embedding(ctx_weights, ctx_compute, input_tensor, layer)
}
\description{
Build embedding forward pass
}
\keyword{internal}
