% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_layers.R
\name{nn_build_flatten}
\alias{nn_build_flatten}
\title{Build flatten forward pass}
\usage{
nn_build_flatten(ctx, input_tensor, layer)
}
\value{
A 2-D \code{ggml_tensor} of shape \code{[features, batch]}.
}
\description{
Build flatten forward pass
}
\keyword{internal}
