% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_model.R
\name{nn_build_graph}
\alias{nn_build_graph}
\title{Build computation graph with allocated weights and inputs}
\usage{
nn_build_graph(model, batch_size, training = TRUE, logits_output = FALSE)
}
\arguments{
\item{model}{Compiled model}

\item{batch_size}{Batch size}
}
\value{
List with ctx_weights, ctx_compute, inputs, outputs, buffer
}
\description{
Build computation graph with allocated weights and inputs
}
\keyword{internal}
