% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_layers.R
\name{nn_build_layer_norm}
\alias{nn_build_layer_norm}
\title{Build layer normalization forward pass}
\usage{
nn_build_layer_norm(ctx, input_tensor, layer)
}
\value{
A \code{ggml_tensor}.
}
\description{
Build layer normalization forward pass
}
\keyword{internal}
