% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_layers.R
\name{nn_build_lstm}
\alias{nn_build_lstm}
\title{Build LSTM forward pass for Sequential model}
\usage{
nn_build_lstm(ctx, input_tensor, layer, batch_size)
}
\description{
Build LSTM forward pass for Sequential model
}
\keyword{internal}
