% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_layers.R
\name{nn_build_max_pooling_2d}
\alias{nn_build_max_pooling_2d}
\title{Build max_pooling_2d forward pass}
\usage{
nn_build_max_pooling_2d(ctx, input_tensor, layer)
}
\value{
A \code{ggml_tensor} holding the 2-D max-pooled output.
}
\description{
Build max_pooling_2d forward pass
}
\keyword{internal}
