% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_model.R
\name{nn_count_layer_params}
\alias{nn_count_layer_params}
\title{Count parameters for a single layer}
\usage{
nn_count_layer_params(layer)
}
\arguments{
\item{layer}{A layer list}
}
\value{
Number of parameters
}
\description{
Count parameters for a single layer
}
\keyword{internal}
