% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_functional.R
\name{nn_functional_output_shape}
\alias{nn_functional_output_shape}
\title{Infer output shape of a functional node given its parent shapes}
\usage{
nn_functional_output_shape(node, parent_shapes)
}
\description{
Infer output shape of a functional node given its parent shapes
}
\keyword{internal}
