% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_functional.R
\name{nn_head_loss}
\alias{nn_head_loss}
\title{Topologically sort nodes reachable from output nodes}
\usage{
nn_head_loss(loss_name, preds_mat, y)
}
\arguments{
\item{outputs}{List of output ggml_tensor_node objects}
}
\value{
Named list: nodes in topological order (inputs first, outputs last)
}
\description{
Topologically sort nodes reachable from output nodes
}
