% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_layers.R
\name{nn_infer_shapes}
\alias{nn_infer_shapes}
\title{Infer shapes for all layers in model}
\usage{
nn_infer_shapes(model)
}
\arguments{
\item{model}{Sequential model}
}
\value{
Model with shapes filled in
}
\description{
Infer shapes for all layers in model
}
\keyword{internal}
