% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_layers.R
\name{nn_init_glorot_uniform}
\alias{nn_init_glorot_uniform}
\title{Initialize weight tensor with Glorot uniform distribution}
\usage{
nn_init_glorot_uniform(tensor, fan_in, fan_out)
}
\description{
Initialize weight tensor with Glorot uniform distribution
}
\keyword{internal}
