% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_layers.R
\name{nn_init_he_uniform}
\alias{nn_init_he_uniform}
\title{Initialize weight tensor with He uniform distribution}
\usage{
nn_init_he_uniform(tensor, fan_in)
}
\value{
Called for its side effect (writes initial weights into \code{tensor}); invisibly returns \code{NULL}.
}
\description{
Initialize weight tensor with He uniform distribution
}
\keyword{internal}
