% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_layers.R
\name{nn_init_zeros}
\alias{nn_init_zeros}
\title{Initialize bias tensor to zeros}
\usage{
nn_init_zeros(tensor)
}
\value{
Called for its side effect (zero-fills \code{tensor}); invisibly returns \code{NULL}.
}
\description{
Initialize bias tensor to zeros
}
\keyword{internal}
