% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_functional.R
\name{print.ggml_functional_model}
\alias{print.ggml_functional_model}
\title{Print method for ggml_functional_model}
\usage{
\method{print}{ggml_functional_model}(x, ...)
}
\arguments{
\item{x}{A ggml_functional_model object}

\item{...}{Additional arguments (ignored)}
}
\value{
The model object (invisibly).
}
\description{
Print method for ggml_functional_model
}
