% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_model.R
\name{print.ggml_history}
\alias{print.ggml_history}
\title{Print method for ggml_history}
\usage{
\method{print}{ggml_history}(x, ...)
}
\arguments{
\item{x}{A ggml_history object}

\item{...}{Additional arguments (ignored)}
}
\value{
The history object (invisibly).
}
\description{
Print method for ggml_history
}
