% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onnx.R
\name{print.onnx_model}
\alias{print.onnx_model}
\title{Print ONNX model summary}
\usage{
\method{print}{onnx_model}(x, ...)
}
\arguments{
\item{x}{An \code{onnx_model} object.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print ONNX model summary
}
