% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quants.R
\name{quantize_row_mxfp4_ref}
\alias{quantize_row_mxfp4_ref}
\title{Quantize Row Reference (MXFP4)}
\usage{
quantize_row_mxfp4_ref(src_data, n_elements)
}
\arguments{
\item{src_data}{Numeric vector of float values to quantize}

\item{n_elements}{Number of elements to quantize}
}
\value{
Raw vector of quantized data
}
\description{
Basic row-level MXFP4 quantization.
}
\seealso{
Other quantization: 
\code{\link{dequantize_row_iq2_xxs}()},
\code{\link{dequantize_row_mxfp4}()},
\code{\link{dequantize_row_nvfp4}()},
\code{\link{dequantize_row_q1_0}()},
\code{\link{dequantize_row_q2_K}()},
\code{\link{dequantize_row_q4_0}()},
\code{\link{dequantize_row_tq1_0}()},
\code{\link{ggml_quant_block_info}()},
\code{\link{iq2xs_free_impl}()},
\code{\link{iq2xs_init_impl}()},
\code{\link{iq3xs_free_impl}()},
\code{\link{iq3xs_init_impl}()},
\code{\link{quantize_iq2_xxs}()},
\code{\link{quantize_mxfp4}()},
\code{\link{quantize_nvfp4}()},
\code{\link{quantize_q1_0}()},
\code{\link{quantize_q2_K}()},
\code{\link{quantize_q4_0}()},
\code{\link{quantize_row_iq3_xxs_ref}()},
\code{\link{quantize_row_nvfp4_ref}()},
\code{\link{quantize_row_q1_0_ref}()},
\code{\link{quantize_row_q2_K_ref}()},
\code{\link{quantize_row_q4_0_ref}()},
\code{\link{quantize_row_tq1_0_ref}()},
\code{\link{quantize_tq1_0}()}
}
\concept{quantization}
