% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autograd.R
\name{ag_param}
\alias{ag_param}
\title{Create a parameter tensor (gradient tracked)}
\usage{
ag_param(
  data,
  device = .ag_device_state$device,
  dtype = .ag_device_state$dtype
)
}
\arguments{
\item{data}{Numeric matrix or vector}

\item{device}{\code{"cpu"} (default) or \code{"gpu"}}
}
\value{
An ag_tensor with requires_grad = TRUE
}
\description{
Create a parameter tensor (gradient tracked)
}
