% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{ggml_graph_reset}
\alias{ggml_graph_reset}
\title{Reset Graph (for backpropagation)}
\usage{
ggml_graph_reset(graph)
}
\arguments{
\item{graph}{Graph object with gradients allocated}
}
\value{
No return value, called for side effects
}
\description{
Resets the computation graph for a new backward pass.
NOTE: This function requires the graph to have gradients allocated
(used for training/backpropagation). For inference-only graphs,
this function will cause an error.
}
