% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{ggml_graph_view}
\alias{ggml_graph_view}
\title{Create a View of a Subgraph}
\usage{
ggml_graph_view(graph, i0, i1)
}
\arguments{
\item{graph}{External pointer to computation graph}

\item{i0}{Start index (0-based, inclusive)}

\item{i1}{End index (exclusive)}
}
\value{
External pointer to graph view
}
\description{
Creates a view of a portion of a computation graph, containing nodes
from index i0 to i1 (exclusive). The view shares the underlying nodes
but does not include leaf tensors or gradients.
}
\examples{
\dontrun{
# Get first 10 nodes of a graph
view <- ggml_graph_view(graph, 0, 10)
}
}
\seealso{
Other graph: 
\code{\link{ggml_op_can_inplace}()}
}
\concept{graph}
