% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{ggml_log}
\alias{ggml_log}
\title{Natural Logarithm (Graph)}
\usage{
ggml_log(ctx, a)
}
\arguments{
\item{ctx}{GGML context}

\item{a}{Input tensor}
}
\value{
Tensor representing the log operation
}
\description{
Creates a graph node for element-wise natural logarithm: log(x)
}
