% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer.R
\name{ggml_opt_set_lr}
\alias{ggml_opt_set_lr}
\title{Set learning rate in optimizer context}
\usage{
ggml_opt_set_lr(lr_ud, adamw_lr = NA, sgd_lr = NA)
}
\arguments{
\item{lr_ud}{LR userdata pointer (from `ggml_opt_init_for_fit()$lr_ud`)}

\item{adamw_lr}{New AdamW learning rate (NA to keep current)}

\item{sgd_lr}{New SGD learning rate (NA to keep current)}
}
\value{
NULL invisibly
}
\description{
Updates the LR used for subsequent backward passes.
Can be called between epochs to implement LR scheduling.
}
\seealso{
Other optimization: 
\code{\link{ggml_fit_opt}()},
\code{\link{ggml_fit_opt_multi}()},
\code{\link{ggml_opt_alloc}()},
\code{\link{ggml_opt_context_optimizer_type}()},
\code{\link{ggml_opt_dataset_data}()},
\code{\link{ggml_opt_dataset_free}()},
\code{\link{ggml_opt_dataset_get_batch}()},
\code{\link{ggml_opt_dataset_get_batch_head}()},
\code{\link{ggml_opt_dataset_init}()},
\code{\link{ggml_opt_dataset_labels}()},
\code{\link{ggml_opt_dataset_ndata}()},
\code{\link{ggml_opt_dataset_shuffle}()},
\code{\link{ggml_opt_dataset_weights}()},
\code{\link{ggml_opt_default_params}()},
\code{\link{ggml_opt_epoch}()},
\code{\link{ggml_opt_eval}()},
\code{\link{ggml_opt_fit}()},
\code{\link{ggml_opt_fit_multi}()},
\code{\link{ggml_opt_free}()},
\code{\link{ggml_opt_get_lr}()},
\code{\link{ggml_opt_grad_acc}()},
\code{\link{ggml_opt_init}()},
\code{\link{ggml_opt_init_for_fit}()},
\code{\link{ggml_opt_init_for_fit_multi}()},
\code{\link{ggml_opt_inputs}()},
\code{\link{ggml_opt_labels}()},
\code{\link{ggml_opt_loss}()},
\code{\link{ggml_opt_loss_i}()},
\code{\link{ggml_opt_loss_type_binary_cross_entropy}()},
\code{\link{ggml_opt_loss_type_cross_entropy}()},
\code{\link{ggml_opt_loss_type_huber}()},
\code{\link{ggml_opt_loss_type_mae}()},
\code{\link{ggml_opt_loss_type_mean}()},
\code{\link{ggml_opt_loss_type_mse}()},
\code{\link{ggml_opt_loss_type_sum}()},
\code{\link{ggml_opt_loss_type_weighted_mse}()},
\code{\link{ggml_opt_n_loss}()},
\code{\link{ggml_opt_ncorrect}()},
\code{\link{ggml_opt_optimizer_name}()},
\code{\link{ggml_opt_optimizer_type_adamw}()},
\code{\link{ggml_opt_optimizer_type_sgd}()},
\code{\link{ggml_opt_outputs}()},
\code{\link{ggml_opt_outputs_i}()},
\code{\link{ggml_opt_pred}()},
\code{\link{ggml_opt_prepare_alloc}()},
\code{\link{ggml_opt_reset}()},
\code{\link{ggml_opt_result_accuracy}()},
\code{\link{ggml_opt_result_free}()},
\code{\link{ggml_opt_result_init}()},
\code{\link{ggml_opt_result_loss}()},
\code{\link{ggml_opt_result_loss_i}()},
\code{\link{ggml_opt_result_n_loss}()},
\code{\link{ggml_opt_result_ndata}()},
\code{\link{ggml_opt_result_pred}()},
\code{\link{ggml_opt_result_reset}()},
\code{\link{ggml_opt_static_graphs}()}
}
\concept{optimization}
