% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\docType{data}
\name{ggml_pool_1d}
\alias{ggml_pool_1d}
\alias{GGML_OP_POOL_MAX}
\alias{GGML_OP_POOL_AVG}
\title{1D Pooling (Graph)}
\format{
An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.
}
\usage{
ggml_pool_1d(ctx, a, op, k0, s0 = k0, p0 = 0L)

GGML_OP_POOL_MAX

GGML_OP_POOL_AVG
}
\arguments{
\item{ctx}{GGML context}

\item{a}{Input tensor}

\item{op}{Pool operation: GGML_OP_POOL_MAX (0) or GGML_OP_POOL_AVG (1)}

\item{k0}{Kernel size}

\item{s0}{Stride (default = k0)}

\item{p0}{Padding (default 0)}
}
\value{
Pooled tensor
}
\description{
Applies 1D pooling operation.
}
\keyword{datasets}
