% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{ggml_soft_max_inplace}
\alias{ggml_soft_max_inplace}
\title{Softmax In-place (Graph)}
\usage{
ggml_soft_max_inplace(ctx, a)
}
\arguments{
\item{ctx}{GGML context}

\item{a}{Input tensor (will be modified in-place)}
}
\value{
View of input tensor with softmax applied
}
\description{
Creates a graph node for in-place softmax operation.
Returns a view of the input tensor.
}
