% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{ggml_view_tensor}
\alias{ggml_view_tensor}
\title{View Tensor}
\usage{
ggml_view_tensor(ctx, src)
}
\arguments{
\item{ctx}{GGML context}

\item{src}{Source tensor}
}
\value{
View tensor (shares data with src)
}
\description{
Creates a view of the tensor (shares data, no copy)
}
