% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vulkan.R
\name{ggml_vulkan_status}
\alias{ggml_vulkan_status}
\title{Print Vulkan status}
\usage{
ggml_vulkan_status()
}
\value{
NULL (invisible), prints status to console
}
\description{
Prints information about Vulkan availability and devices.
}
\examples{
ggml_vulkan_status()
}
