% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggml.R
\docType{package}
\name{ggmlR-package}
\alias{ggmlR}
\alias{ggmlR-package}
\title{ggmlR: GGML Tensor Operations for Machine Learning}
\description{
An R port of the GGML library providing efficient tensor operations for machine learning on CPU. Implements core functionality of GGML, including element-wise arithmetic and basic tensor manipulation. The package brings the lightweight and performance-oriented design of GGML to the R environment for fast numerical computation and model prototyping.
}
\author{
\strong{Maintainer}: Yuri Baramykov \email{lbsbmsu@mail.ru}

}
\keyword{internal}
