% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gguf.R
\name{gguf_free}
\alias{gguf_free}
\title{Free GGUF Resources}
\usage{
gguf_free(x)
}
\arguments{
\item{x}{A \code{gguf} object.}
}
\description{
Explicitly frees the internal GGUF context. Called automatically by the
garbage collector, but can be called manually to release memory sooner.
}
