% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_layers.R
\name{nn_build_batch_norm}
\alias{nn_build_batch_norm}
\title{Build batch_norm forward pass}
\usage{
nn_build_batch_norm(ctx, input_tensor, layer)
}
\description{
Build batch_norm forward pass
}
\keyword{internal}
