% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_layers.R
\name{nn_build_gru}
\alias{nn_build_gru}
\title{Build GRU forward pass for Sequential model}
\usage{
nn_build_gru(ctx, input_tensor, layer, batch_size)
}
\description{
Build GRU forward pass for Sequential model
}
\keyword{internal}
