% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_layers.R
\name{nn_build_layer}
\alias{nn_build_layer}
\title{Build a layer's forward pass}
\usage{
nn_build_layer(ctx, input_tensor, layer, training = TRUE, ctx_weights = NULL)
}
\description{
Build a layer's forward pass
}
\keyword{internal}
