% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\docType{data}
\name{rope_types}
\alias{rope_types}
\alias{GGML_ROPE_TYPE_NORM}
\alias{GGML_ROPE_TYPE_NEOX}
\alias{GGML_ROPE_TYPE_MROPE}
\alias{GGML_ROPE_TYPE_VISION}
\title{RoPE Mode Constants}
\format{
An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.
}
\usage{
GGML_ROPE_TYPE_NORM

GGML_ROPE_TYPE_NEOX

GGML_ROPE_TYPE_MROPE

GGML_ROPE_TYPE_VISION
}
\description{
Constants for RoPE (Rotary Position Embedding) modes.
}
\details{
- GGML_ROPE_TYPE_NORM (0): Standard RoPE as in original paper
- GGML_ROPE_TYPE_NEOX (2): GPT-NeoX style RoPE (interleaved differently)
- GGML_ROPE_TYPE_MROPE (8): Multi-RoPE for models like Qwen2-VL
- GGML_ROPE_TYPE_VISION (24): Vision model RoPE
}
\keyword{datasets}
